refactor: change homepage to 我的庫存
#33
